+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_updater/impl.ex
@@ -525,6 +525,19 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
     now_secs + ttl + 3600 < expiry_secs
   end
 
+  def validate_value_type(expected_type, %Bson.UTC{} = value) do
+    ValueType.validate_value(expected_type, value)
+  end
+
+  def validate_value_type(expected_type, %Bson.Bin{} = value) do
+    ValueType.validate_value(expected_type, value)
+  end
+
+  # Explicitly match on all structs to avoid pattern matching them as maps below
+  def validate_value_type(_expected_type, %_{} = _unsupported_struct) do
+    {:error, :unexpected_value_type}
+  end
+
   def validate_value_type(expected_types, %{} = object) do
     Enum.reduce_while(object, :ok, fn {path, value}, _acc ->
       key = Atom.to_string(path)
@@ -1118,10 +1131,14 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
   end
 
   defp forget_interfaces(state, interfaces_to_drop) do
-    updated_triggers =
-      Enum.reduce(interfaces_to_drop, state.data_triggers, fn iface, data_triggers ->
-        interface_id = Map.fetch!(state.interfaces, iface).interface_id
+    iface_ids_to_drop =
+      Enum.filter(interfaces_to_drop, &Map.has_key?(state.interfaces, &1))
+      |> Enum.map(fn iface ->
+        Map.fetch!(state.interfaces, iface).interface_id
+      end)
 
+    updated_triggers =
+      Enum.reduce(iface_ids_to_drop, state.data_triggers, fn interface_id, data_triggers ->
         Enum.reject(data_triggers, fn {{_event_type, iface_id, _endpoint}, _val} ->
           iface_id == interface_id
         end)
@@ -1129,9 +1146,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
       end)
 
     updated_mappings =
-      Enum.reduce(interfaces_to_drop, state.mappings, fn iface, mappings ->
-        interface_id = Map.fetch!(state.interfaces, iface).interface_id
-
+      Enum.reduce(iface_ids_to_drop, state.mappings, fn interface_id, mappings ->
         Enum.reject(mappings, fn {_endpoint_id, mapping} ->
           mapping.interface_id == interface_id
         end)
@@ -1139,8 +1154,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Impl do
       end)
 
     updated_ids =
-      Enum.reduce(interfaces_to_drop, state.interface_ids_to_name, fn iface, ids ->
-        interface_id = Map.fetch!(state.interfaces, iface).interface_id
+      Enum.reduce(iface_ids_to_drop, state.interface_ids_to_name, fn interface_id, ids ->
         Map.delete(ids, interface_id)
       end)