+++ apps/astarte_data_updater_plant/CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.10.0-beta.3] - Unreleased
+## [0.10.0-rc.0] - Unreleased
+
+## [0.10.0-beta.3] - 2018-12-19
 ### Fixed
 - Correctly handle Bson.UTC and Bson.Bin incoming data.
 - Fix crash when an interface that has been previously removed from the device introspection expires from cache.